Movded mask implementation to model, as it does not belong to server pkg
